Do not put a rest at the end of line based on beat
Sometimes putting the rest at the end of a line is better and sometimes
it's not. Deciding that based on the beat isn't always correct.
Therefore remove the mechanism from the code.

Now the default behavior is to put the rest at the beginning of a line.
Users can use the special symbol "O" to manually control the position of
each rest.