Support <group_8th_notes>
Add support for "<group_8th_notes> VALUE" in melody.txt. When VALUE is 0
(default), underlines can only extend up to 1 beat (for 2/4 or 4/4 time
signatures). When VALUE is 1, the underlines are allowed to extend to 2
beats, provided that the underline group only contains 8th notes and
nothing else.

Consider the following example:

 <time> 4/4 hyphen=8
 1122 | 3-3-4-4-

If VALUE is 0, the 4 notes in the first bar will be divided into 2
groups, each containing 2 notes. If VALUE is 1, the 4 notes will be put
into 1 group.

Consider another example:

 <time> 4/4 hyphen=8
 111 | 3-3-4-4-

If VALUE is 1, then the 3 notes in the first bar will be put into 1
group.