Decrease vertical space between upper octave dot and curve
If a triplet on the bottom line of a page starts from (or ends with) a
note with an upper octave dot, the bottom part of the page would be too
high, causing it to be moved to the next page. To solve the problem,
decrease the vertical space from an upper dot to a curve.